requirements.txt
# Flask and extensions
Flask==2.3.3
Werkzeug==2.3.7

# Database
psycopg2-binary==2.9.9

# Email
sendgrid==6.10.0

# Time zone handling
pytz==2024.1

# HTTP requests
requests==2.31.0

# Geolocation
geopy==2.4.1

# URL parsing and handling
urllib3==2.1.0

# Date and time handling
python-dateutil==2.8.2

# WSGI HTTP Server
gunicorn==21.2.0

# Environment variables
python-dotenv==1.0.0
